config: Extract `RateLimitsConfig` struct
This moves the `rate_limiter` map and the `new_version_rate_limit` setting
from `Server` onto a dedicated `RateLimitsConfig` struct in
`config::rate_limits`, along with the per-action loading loop. The fields
are renamed to `actions` and `new_versions_daily` since the struct now
provides the `rate_limits` namespace.
