+++ src/tests/routes/crates/new.rs
@@ -6,7 +6,12 @@ use insta::assert_snapshot;
 async fn daily_limit() {
     let (app, _, user) = TestApp::full().with_user().await;
 
-    let max_daily_versions = app.as_inner().config.new_version_rate_limit.unwrap();
+    let max_daily_versions = app
+        .as_inner()
+        .config
+        .rate_limits
+        .new_versions_daily
+        .unwrap();
     for version in 1..=max_daily_versions {
         let crate_to_publish = PublishBuilder::new("foo_daily_limit", &format!("0.0.{version}"));
         user.publish_crate(crate_to_publish).await.good();
